REmove unnecesary files #16 Remove Resources Move Filtro
