+++ src/party.windows/party.windows.csproj
@@ -7,15 +7,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="runtimes\**" />
-    <EmbeddedResource Remove="runtimes\**" />
-    <None Remove="runtimes\**" />
-  </ItemGroup>
-
-  <ItemGroup>
     <None Remove="appsettings.json" />
     <None Remove="images\Background.jpg" />
-    <None Remove="Fonts\OFL.txt" />
+    <None Remove="fonts\OFL.txt" />
     <None Remove="images\Icon.ico" />
     <None Remove="images\barcode1.png" />
     <None Remove="images\barcode2.png" />
@@ -26,7 +20,7 @@
     <Content Include="images\Background.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Content Include="Fonts\OFL.txt" />
+    <Content Include="fonts\OFL.txt" />
     <Content Include="images\Icon.ico">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
@@ -64,18 +58,7 @@
     <Compile Update="forms\SettingsForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Update="Properties\Resources.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Resources.resx</DependentUpon>
-    </Compile>
   </ItemGroup>
 
-  <ItemGroup>
-    <EmbeddedResource Update="Properties\Resources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-  </ItemGroup>
 
 </Project>
